Renamed cloud_api.py to cloud_api_iot.py
